# api/internal/db/sqlc_models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type AdminSession struct {
	TokenHash string
	UserID    int64
	CreatedAt pgtype.Timestamptz
	ExpiresAt pgtype.Timestamptz
}

type AdminUser struct {
	ID           int64
	Username     string
	PasswordHash string
	DisplayName  string
	CreatedAt    pgtype.Timestamptz
}

type Order struct {
	ID             int64
	IdempotencyKey string
	CustomerName   string
	Phone          string
	Comment        string
	Items          []byte
	Total          int32
	CreatedAt      pgtype.Timestamptz
}

type Outbox struct {
	ID          int64
	Kind        string
	Payload     []byte
	Status      string
	Attempts    int32
	NextRetryAt pgtype.Timestamptz
	LastError   string
	CreatedAt   pgtype.Timestamptz
	UpdatedAt   pgtype.Timestamptz
}

type Product struct {
	ID        int64
	Slug      string
	Brand     string
	Model     string
	Name      string
	SizeLabel string
	Width     int32
	Profile   int32
	Diameter  int32
	Season    string
	Spikes    bool
	Runflat   bool
	Price     int32
	Stock     int32
	SyncedAt  pgtype.Timestamptz
	Code      string
	ImageUrl  string
}

type ProductOverride struct {
	Slug      string
	Hidden    bool
	BadgeHit  bool
	UpdatedAt pgtype.Timestamptz
	UpdatedBy *int64
}