# api/openapi.yaml
openapi: 3.0.3
info:
  title: TireStock API
  description: >
    Контракт витрины TireStock. Источник истины: любое изменение API начинается
    с правки этого файла, из него генерируются TS-типы в web/lib/api/.
  version: 0.1.0

servers:
  - url: /api/v1

paths:
  /health:
    get:
      operationId: getHealth
      summary: Проверка живости сервиса и БД
      responses:
        "200":
          description: Сервис и БД в порядке
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [ok]
        "503":
          $ref: "#/components/responses/Error"

  /products:
    get:
      operationId: listProducts
      summary: Каталог шин с фильтрами и пагинацией
      description: >
        Имена фильтров совпадают с query-параметрами URL витрины.
      parameters:
        - { name: q,         in: query, schema: { type: string }, description: "Текстовый поиск по названию/бренду/модели" }
        - { name: width,     in: query, schema: { type: integer }, description: "Ширина, мм (205)" }
        - { name: profile,   in: query, schema: { type: integer }, description: "Профиль, % (55)" }
        - { name: diameter,  in: query, schema: { type: integer }, description: "Диаметр, дюймы (16)" }
        - { name: season,    in: query, schema: { $ref: "#/components/schemas/Season" } }
        - { name: brand,     in: query, schema: { type: string }, description: "Бренд (Nokian)" }
        - { name: price_min, in: query, schema: { type: integer, minimum: 0 }, description: "Цена от, ₽" }
        - { name: price_max, in: query, schema: { type: integer, minimum: 0 }, description: "Цена до, ₽" }
        - { name: spikes,    in: query, schema: { type: boolean }, description: "Шипы" }
        - { name: runflat,   in: query, schema: { type: boolean }, description: "RunFlat" }
        - { name: page,      in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: per_page,  in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 24 } }
      responses:
        "200":
          description: Страница каталога
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
        "400":
          $ref: "#/components/responses/Error"

  /products/{slug}:
    get:
      operationId: getProductBySlug
      summary: Товар по slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Товар
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/Error"

  /orders:
    post:
      operationId: createOrder
      summary: Оформление заказа
      description: >
        Заказ пишется в БД транзакционно со строкой outbox; воркер доставляет
        его в tradesk с ретраями — заказ никогда не теряется. Idempotency-key
        (заголовок или поле тела) защищает от даблклика «Оформить заказ».
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Заказ принят
          content:
            application/json:
              schema:
                type: object
                required: [order_id]
                properties:
                  order_id:
                    type: integer
                    format: int64
        "400":
          $ref: "#/components/responses/Error"

  /callbacks:
    post:
      operationId: createCallback
      summary: Заявка «обратный звонок»
      description: Попадает в tradesk (раздел «Обратный звонок») через outbox.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallbackInput"
      responses:
        "201":
          description: Заявка принята
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [accepted]
        "400":
          $ref: "#/components/responses/Error"

components:
  responses:
    Error:
      description: Ошибка в едином формате
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"

  schemas:
    Season:
      type: string
      enum: [summer, winter, allseason]

    Product:
      type: object
      required:
        [id, slug, brand, model, name, size_label, width, profile, diameter,
         season, spikes, runflat, price, stock, image_url, badge_hit]
      properties:
        id: { type: integer, format: int64 }
        slug: { type: string, example: nokian-hakkapeliitta-10p-205-55-r16 }
        brand: { type: string, example: Nokian }
        model: { type: string, example: Hakkapeliitta 10p }
        name: { type: string, example: Nokian Hakkapeliitta 10p 205/55 R16 }
        size_label: { type: string, example: 205/55 R16 94T }
        width: { type: integer, example: 205 }
        profile: { type: integer, example: 55 }
        diameter: { type: integer, example: 16 }
        season: { $ref: "#/components/schemas/Season" }
        spikes: { type: boolean }
        runflat: { type: boolean }
        price: { type: integer, description: "Рубли за штуку", example: 12490 }
        stock: { type: integer, example: 16 }
        image_url: { type: string, description: "Фото из SelectTyres (пусто → плейсхолдер)" }
        badge_hit: { type: boolean, description: "Бейдж «Хит» (оверрайд админки)" }

    ProductList:
      type: object
      required: [items, total, page, per_page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }

    OrderItem:
      type: object
      required: [slug, name, price, qty]
      properties:
        slug: { type: string }
        name: { type: string }
        price: { type: integer, description: "Рубли за штуку" }
        qty: { type: integer, minimum: 1, default: 4, description: "Шины покупают комплектами — дефолт 4" }

    CreateOrderInput:
      type: object
      required: [customer_name, phone, items]
      properties:
        idempotency_key: { type: string }
        customer_name: { type: string, example: Иван Петров }
        phone: { type: string, example: "+7 (921) 123-45-67" }
        comment: { type: string }
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItem"

    CallbackInput:
      type: object
      required: [phone]
      properties:
        name: { type: string, example: Иван }
        phone: { type: string, example: "+7 (921) 123-45-67" }
        comment: { type: string }

    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: "Стабильный строковый код: validation_failed, not_found, internal_error, db_unavailable"
            message: { type: string }